Migrate mapper to use less context

+++ shared/src/commonMain/kotlin/fr/outadoc/justchatting/component/twitch/websocket/irc/model/IrcEvent.kt
@@ -42,7 +42,7 @@ sealed interface IrcEvent {
             data class PaidMessageInfo(
                 val amount: Long,
                 val currency: String,
-                val exponent: Int,
+                val exponent: Long,
                 val isSystemMessage: Boolean,
                 val level: String,
             )
